20221205 - Ziyu Wang 1. Fixed log2 overflow in Tile constructor 2. Implement a result counter for termination To Do: 1. Implement System Configuration Output 2. Implement OpenCV compiling 3. Change some pointer to reference

+++ tile.h
@@ -34,7 +34,7 @@ private:
     size_t arrayNumY_;
 
     // Latency configuration
-    int latencyVMM_;
+    size_t latencyVMM_;
     const static int concHTree_ = 1;
     const static int addHTree_ = 1;
